Extract interface for SubscriptionResource
